Datasets data.ex in TAM Package

Datasets included in the TAM package

data(data.ex08)
data(data.ex10)
data(data.ex11)
data(data.ex12)
data(data.ex14)
data(data.ex15)
data(data.exJ03)

Format

Data data.ex08 for Example 8 in tam.mml has the following format:

List of 2
$ facets:'data.frame': 1000 obs. of 1 variable:
..$ female: int [1:1000] 1 1 1 1 1 1 1 1 1 1 ...
$ resp : num [1:1000, 1:10] 1 1 1 0 1 0 1 1 0 1 ...
..- attr(*, "dimnames")=List of 2
.. ..$ : NULL
.. ..$ : chr [1:10] "I0001" "I0002" "I0003" "I0004" ...

Data data.ex10 for Example 10 in tam.mml has the following format:

'data.frame': 675 obs. of 7 variables:
$ pid : int 1 1 1 2 2 3 3 4 4 5 ...
$ rater: int 1 2 3 2 3 1 2 1 3 1 ...
$ I0001: num 0 1 1 1 1 1 1 1 1 1 ...
$ I0002: num 1 1 1 1 1 0 1 1 1 1 ...
$ I0003: num 1 1 1 1 0 0 0 1 0 1 ...
$ I0004: num 0 1 0 0 1 0 1 0 1 0 ...
$ I0005: num 0 0 1 1 1 0 0 1 0 1 ...

Data data.ex11 for Example 11 in tam.mml has the following format:

'data.frame': 3400 obs. of 13 variables:
$ booklet: chr "B1" "B1" "B3" "B2" ...
$ M133 : int 1 1 NA 1 NA 1 NA 1 0 1 ...
$ M176 : int 1 0 1 NA 0 0 0 NA NA NA ...
$ M202 : int NA NA NA 0 NA NA NA 0 0 0 ...
$ M212 : int NA NA 1 0 0 NA 0 1 0 0 ...
$ M214 : int 1 0 1 1 0 0 0 0 1 0 ...
$ M259 : int NA NA 1 1 1 NA 1 1 1 1 ...
$ M303 : int NA NA 1 1 1 NA 1 1 1 0 ...
$ M353 : int NA NA NA 1 NA NA NA 1 1 9 ...
$ M355 : int NA NA NA 1 NA NA NA 1 1 0 ...
$ M444 : int 0 0 0 NA 0 0 0 NA NA NA ...
$ M446 : int 1 0 0 1 0 1 1 1 0 0 ...
$ M449 : int NA NA NA 1 NA NA NA 1 1 1 ...

Missing responses by design are coded as NA, omitted responses are coded as 9.

Data data.ex12 for Example 12 in tam.mml has the following format:

num [1:100, 1:10] 1 1 1 1 1 1 1 1 1 1 ...
- attr(*, "dimnames")=List of 2
..$ : NULL
..$ : chr [1:10] "I0001" "I0002" "I0003" "I0004" ...

Data data.ex14 for Example 14 in tam.mml has the following format:

'data.frame': 1110 obs. of 11 variables:
$ pid : num 1001 1001 1001 1001 1001 ...
$ X1 : num 1 1 1 1 1 1 0 0 0 0 ...
$ X2 : int 1 1 1 1 1 1 1 1 1 1 ...
$ rater: int 4 4 4 4 4 4 4 4 4 4 ...
$ crit1: int 0 0 2 1 1 2 0 0 0 0 ...
$ crit2: int 0 0 0 0 0 0 0 0 0 0 ...
$ crit3: int 0 1 1 0 0 1 0 0 1 0 ...
$ crit4: int 0 0 0 1 0 0 0 0 0 0 ...
$ crit5: int 0 0 0 0 1 1 0 0 0 0 ...
$ crit6: int 0 0 0 0 1 0 0 0 0 0 ...
$ crit7: int 1 0 2 0 0 0 0 0 0 0 ...

Data data.ex15 for Example 15 in tam.mml has the following format:

'data.frame': 2155 obs. of 182 variables:
$ pid : num 10001 10002 10003 10004 10005 ...
$ group : num 1 1 0 0 1 0 1 0 1 1 ...
$ Item001: num 0 NA NA 0 NA NA NA 0 0 NA ...
$ Item002: num 1 NA NA 1 NA NA NA NA 1 NA ...
$ Item003: num NA NA NA NA 1 NA NA NA NA 1 ...
$ Item004: num NA NA 0 NA NA NA NA NA NA NA ...
$ Item005: num NA NA 1 NA NA NA NA NA NA NA ...
[...]

This dataset shows an atypical convergence behavior. Look at Example 15 to fix convergence problems using arguments increment.factor and fac.oldxsi.

Data data.exJ03 for Example 4 in tam.jml has the following format:

List of 2
$ resp:'data.frame': 40 obs. of 20 variables:
..$ I104: int [1:40] 4 5 6 5 3 4 3 5 4 6 ...
..$ I118: int [1:40] 6 4 6 5 3 2 5 3 5 4 ...
[...] ..$ I326: int [1:40] 6 1 5 1 4 2 4 1 6 1 ...
..$ I338: int [1:40] 6 2 6 1 6 2 4 1 6 1 ...
$ X :'data.frame': 40 obs. of 4 variables:
..$ rater : int [1:40] 40 40 96 96 123 123 157 157 164 164 ...
..$ gender: int [1:40] 2 2 1 1 1 1 2 2 2 2 ...
..$ region: num [1:40] 1 1 1 1 2 2 1 1 1 1 ...
..$ leader: int [1:40] 1 2 1 2 1 2 1 2 1 2 ...

It is a rating dataset (a subset of a dataset provided by Matt Barney).

Data data.ex16 contains dichotomous item response data from three studies corresponding to three grades.

'data.frame': 3235 obs. of 25 variables:
$ idstud: num 1e+05 1e+05 1e+05 1e+05 1e+05 ...
$ grade : num 1 1 1 1 1 1 1 1 1 1 ...
$ A1 : int 1 1 1 1 1 1 1 1 1 1 ...
$ B1 : int 1 1 1 1 1 1 1 1 1 1 ...
$ C1 : int 1 1 1 1 1 1 1 1 1 1 ...
$ D1 : int 1 1 1 1 1 1 1 1 1 1 ...
$ E1 : int 0 0 1 0 1 1 1 0 1 1 ...
$ E2 : int 1 1 1 1 1 1 1 0 1 1 ...
$ E3 : int 1 1 1 1 1 1 1 0 1 1 ...
$ F1 : int 1 0 1 1 0 0 1 0 1 1 ...
$ G1 : int 0 1 1 1 1 0 1 0 1 1 ...
$ G2 : int 1 1 1 1 1 0 0 0 1 1 ...
$ G3 : int 1 0 1 1 1 0 0 0 1 1 ...
$ H1 : int 1 0 1 1 1 0 0 0 1 1 ...
$ H2 : int 1 0 1 1 1 0 0 0 1 1 ...
$ I1 : int 1 0 1 0 1 0 0 0 1 1 ...
$ I2 : int 1 0 1 0 1 0 0 0 1 1 ...
$ J1 : int NA NA NA NA NA NA NA NA NA NA ...
$ K1 : int NA NA NA NA NA NA NA NA NA NA ...
$ L1 : int NA NA NA NA NA NA NA NA NA NA ...
$ L2 : int NA NA NA NA NA NA NA NA NA NA ...
$ L3 : int NA NA NA NA NA NA NA NA NA NA ...
$ M1 : int NA NA NA NA NA NA NA NA NA NA ...
$ M2 : int NA NA NA NA NA NA NA NA NA NA ...
$ M3 : int NA NA NA NA NA NA NA NA NA NA ...

Data data.ex17 contains polytomous item response data from three studies corresponding to three grades.

'data.frame': 3235 obs. of 15 variables:
$ idstud: num 1e+05 1e+05 1e+05 1e+05 1e+05 ...
$ grade : num 1 1 1 1 1 1 1 1 1 1 ...
$ A : int 1 1 1 1 1 1 1 1 1 1 ...
$ B : int 1 1 1 1 1 1 1 1 1 1 ...
$ C : int 1 1 1 1 1 1 1 1 1 1 ...
$ D : int 1 1 1 1 1 1 1 1 1 1 ...
$ E : num 2 2 3 2 3 3 3 0 3 3 ...
$ F : int 1 0 1 1 0 0 1 0 1 1 ...
$ G : num 2 2 3 3 3 0 1 0 3 3 ...
$ H : num 2 0 2 2 2 0 0 0 2 2 ...
$ I : num 2 0 2 0 2 0 0 0 2 2 ...
$ J : int NA NA NA NA NA NA NA NA NA NA ...
$ K : int NA NA NA NA NA NA NA NA NA NA ...
$ L : num NA NA NA NA NA NA NA NA NA NA ...
$ M : num NA NA NA NA NA NA NA NA NA NA ...

See also