Residuals in an IRT Model

Defines an S3 method for the computation of observed residual values. The computation of residuals is based on weighted likelihood estimates as person parameters, see tam.wle. IRT.residuals can only be applied for unidimensional IRT models. The methods IRT.residuals and residuals are equivalent.

IRT.residuals(object, ...)

# S3 method for tam.mml
IRT.residuals(object, ...)
# S3 method for tam.mml
residuals(object, ...)

# S3 method for tam.mml.2pl
IRT.residuals(object, ...)
# S3 method for tam.mml.2pl
residuals(object, ...)

# S3 method for tam.mml.mfr
IRT.residuals(object, ...)
# S3 method for tam.mml.mfr
residuals(object, ...)

# S3 method for tam.jml
IRT.residuals(object, ...)
# S3 method for tam.jml
residuals(object, ...)

Arguments

object
Object of class tam.mml, tam.mml.2pl or tam.mml.mfr.

...
Further arguments to be passed

Value

List with following entries

residuals

Residuals

stand_residuals

Standardized residuals

X_exp

Expected value of the item response \(X_{pi}\)

X_var

Variance of the item response \(X_{pi}\)

theta

Used person parameter estimate

probs

Expected item response probabilities

Note

Residuals can be used to inspect local dependencies in the item response data, for example using principle component analysis or factor analysis (see Example 1).

See also

Examples

#############################################################################
# EXAMPLE 1: Residuals data.read
#############################################################################

library(sirt)
data(data.read,package="sirt")
dat <- data.read

# for Rasch model
mod <- TAM::tam.mml( dat )
# extract residuals
res <- TAM::IRT.residuals( mod )
str(res)