Utility Functions in TAM

Utility functions in TAM.

## RISE item fit statistic of two models
IRT.RISE( mod_p, mod_np, use_probs=TRUE )

## information about used package version
tam_packageinfo(pack)
## call statement in a string format
tam_print_call(CALL)
## information about R session
tam_rsessinfo()
## grep list of arguments for a specific variable
tam_args_CALL_search(args_CALL, variable, default_value)
## requireNamespace with message of needed installation
require_namespace_msg(pkg)
## add leading zeroes
add.lead(x, width=max(nchar(x)))
## round some columns in a data frame
tam_round_data_frame(obji, from=1, to=ncol(obji), digits=3, rownames_null=FALSE)
## round some columns in a data frame and print this data frame
tam_round_data_frame_print(obji, from=1, to=ncol(obji), digits=3, rownames_null=FALSE)
## copy of CDM::osink
tam_osink(file, suffix="__SUMMARY.Rout")
## copy of CDM::csink
tam_csink(file)

## base::matrix function with argument value byrow=TRUE
tam_matrix2(x, nrow=NULL, ncol=NULL)
## more efficient base::outer functions for operations "*", "+" and "-"
tam_outer(x, y, op="*")
## row normalization of a matrix
tam_normalize_matrix_rows(x)
## row normalization of a vector
tam_normalize_vector(x)
## aggregate function for mean and sum based on base::rowsum
tam_aggregate(x, group, mean=FALSE, na.rm=TRUE)
## column index when a value in a matrix is exceeded (used in TAM::tam.pv)
tam_interval_index(matr, rn)
## cumulative sum of row entries in a matrix
tam_rowCumsums(matr)
## extension of mvtnorm::dmvnorm to matrix entries of mean
tam_dmvnorm(x, mean, sigma, log=FALSE )
## Bayesian bootstrap in TAM (used in tam.pv.mcmc)
tam_bayesian_bootstrap(N, sample_integers=FALSE, do_boot=TRUE)
## weighted covariance matrix
tam_cov_wt(x, wt=NULL, method="ML")
## weighted correlation matrix
tam_cor_wt(x, wt=NULL, method="ML")
## generalized inverse
tam_ginv(x, eps=.05)
## generalized inverse with scaled matrix using MASS::ginv
tam_ginv_scaled(x, use_MASS=TRUE)

## remove items or persons with complete missing entries
tam_remove_missings( dat, items, elim_items=TRUE, elim_persons=TRUE )
## compute AXsi given A and xsi
tam_AXsi_compute(A, xsi)
## fit xsi given A and AXsi
tam_AXsi_fit(A, AXsi)

## maximum absolute difference between objects
tam_max_abs( list1, list2, label )
tam_max_abs_list( list1, list2)

## trimming increments in iterations
tam_trim_increment(increment, max.increment, trim_increment="cut",
     trim_incr_factor=2, eps=1E-10, avoid_na=FALSE)
## numerical differentiation by central difference
tam_difference_quotient(d0, d0p, d0m, h)
## assign elements of a list in an environment
tam_assign_list_elements(x, envir)

Arguments

mod_p
Fitted model

mod_np
Fitted model

use_probs
Logical

pack
An R package

CALL
An R call

args_CALL
Arguments obtained from as.list( sys.call() )

variable
Name of a variable

default_value
Default value of a variable

pkg
String

x
Vector or matrix or list

width
Number of zeroes before decimal

obji
Data frame or vector

from
Integer

to
Integer

digits
Integer

rownames_null
Logical

file
File name

suffix
Suffix for file name of summary output

nrow
Number of rows

ncol
Number of columns

y
Vector

op
An operation "*", "+" or "-"

group
Vector of grouping identifiers

mean
Logical indicating whether mean should be calculated or the sum or vector or matrix

na.rm
Logical indicating whether missing values should be removed

matr
Matrix

sigma
Matrix

log
Logical

N
Integer

sample_integers
Logical indicating whether weights for complete cases should be sampled in bootstrap

do_boot
Logical

wt
Optional vector containing weights

method
Method, see stats::cov.wt

rn
Vector

dat
Data frame

items
Vector

elim_items
Logical

elim_persons
Logical

A
Array

xsi
Vector

AXsi
Matrix

increment
Vector

max.increment
Numeric

trim_increment
One of the methods "half" or "cut"

trim_incr_factor
Factor of trimming in method "half"

eps
Small number preventing from division by zero

use_MASS
Logical indicating whether MASS package should be used.

avoid_na
Logical indicating whether missing values should be set to zero.

d0
Vector

d0p
Vector

d0m
Vector

h
Vector

envir
Environment variable

list1
List

list2
List

label
Element of a list