Joint Maximum Likelihood Estimation

This function estimate unidimensional item response models with joint maximum likelihood (JML, see e.g. Linacre, 1994).

tam.jml(resp, group=NULL, adj=.3, disattenuate=FALSE, bias=TRUE,
    xsi.fixed=NULL, xsi.inits=NULL, theta.fixed=NULL, A=NULL, B=NULL, Q=NULL,
    ndim=1, pweights=NULL, constraint="cases", verbose=TRUE, control=list(), version=3,
    theta_proc=NULL)

# S3 method for tam.jml
summary(object, file=NULL, ...)

# S3 method for tam.jml
logLik(object, ...)

Arguments

resp

A matrix of item responses. Missing responses must be declared as NA.

group

An optional vector of group identifier

disattenuate

An optional logical indicating whether the person parameters should be disattenuated due to unreliability? The disattenuation is conducted by applying the Kelley formula.

adj

Adjustment constant which is subtracted or added to extreme scores (score of zero or maximum score). The default is a value of 0.3

bias

A logical which indicates if JML bias should be reduced by multiplying item parameters by the adjustment factor of \((I-1)/I\)

xsi.fixed

An optional matrix with two columns for fixing some of the basis parameters \(\xi\) of item intercepts. 1st column: Index of \(\xi\) parameter, 2nd column: Fixed value of \(\xi\) parameter

xsi.inits

An optional vector of initial \(\xi\) parameters. Note that all parameters must be specified and the vector is not of the same format as xsi.fixed.

theta.fixed

Matrix for fixed person parameters \(\theta\). The first column includes the index whereas the second column includes the fixed value. This argument can only be applied for version=1.

A

A design array \(A\) for item category intercepts. For item \(i\) and category \(k\), the threshold is specified as \( \sum _j a_{ikj} \xi_j\).

B

A design array for scoring item category responses. Entries in \(B\) represent item loadings on abilities \(\theta\).

Q

A Q-matrix which defines loadings of items on dimensions.

ndim

Number of dimensions in the model. The default is 1.

pweights

An optional vector of person weights.

constraint

Type of constraint for means. Either "cases" (set mean of person parameters to zero) or "items" (set mean of item parameters to zero).

verbose

Logical indicating whether output should be printed during iterations. This argument replaces control$progress.

control

A list of control arguments. See tam.mml for more details.

version

Version function which should be used. version=2 is the former tam.jml2 function in TAM (<2.0). The default version=3 allows efficient handling in case of missing data.

theta_proc

Function for processing theta within iterations. Can only be applied for version=1.

object

Object of class tam.jml (only for summary.tam function)

file

A file name in which the summary output will be written (only for summary.tam.jml function)

...

Further arguments to be passed

Value

A list with following entries

item1

Data frame with item parameters

xsi

Vector of item parameters \(\xi\)

errorP

Standard error of item parameters \(\xi\)

theta

MLE in final step

errorWLE

Standard error of WLE

WLE

WLE in last iteration

WLEreliability

WLE reliability

PersonScores

Scores for each person (sufficient statistic)

ItemScore

Sufficient statistic for each item parameter

PersonMax

Maximum person score

ItemMax

Maximum item score

deviance

Deviance

deviance.history

Deviance history in iterations

resp

Original data frame

resp.ind

Response indicator matrix

group

Vector of group identifiers (if provided as an argument)

pweights

Vector of person weights

A

Design matrix \(A\) of item intercepts

B

Loading (or scoring) matrix \(B\)

nitems

Number of items

maxK

Maximum number of categories

nstud

Number of persons in resp

resp.ind.list

Like resp.ind, only in the format of a list

xsi.fixed

Fixed \(\xi\) item parameters

control

Control list

item

Extended data frame of item parameters

theta_summary

Summary of person parameters

...

References

Linacre, J. M. (1994). Many-Facet Rasch Measurement. Chicago: MESA Press.

Note

This joint maximum likelihood estimation procedure should be compatible with Winsteps and Facets software, see also http://www.rasch.org/software.htm.

See also

Examples